Fix two minor bugs per NOVAS website
The NOVAS website here [0] describes the errors. To quote:

"NOVAS C3.1 novas.c: sidereal_time units bug

When NOVAS C3.1 computes Greenwich Mean Sidereal Time (GMST; gst_type = 0)
using the method based on the Celestial Intermediate Origin (CIO; method = 0),
the units for the equation of the equinox are calculated incorrectly.

In sidereal_time, the line

    ha_eq -= (eqeq / 240.0);

should read

    ha_eq -= (eqeq / 3600.0);

since the eqeq variable has been previously converted to arcseconds instead of
seconds in time.  NOVAS C3.1 eph_manager.c: ephem_close does not reset the
EPHFILE pointer

The check to see if an ephemeris file is already open will always be true if
an ephemeris file was open at any point in the program, even if ephem_close
was called. ephem_close does not reset the EPHFILE pointer to NULL. This bug
will be addressed during the next overhaul of the ephemeris-access software.

Memory errors occur when a call to ephem_close occurs between two calls to
ephem_open. The second time ephem_open is called, if (EPHFILE) is TRUE because
EPHFILE is still the address that was assigned the first time.

Two calls to ephem_open without an intervening call to ephem_close will work
because the EPHFILE still is open when if (EPHFILE) checks during the second
ephem_open call.

Users who open multiple ephemeris files within a particular program may wish
to edit the ephem_close function as follows:

    if (EPHFILE)
    {
    error = (short int) fclose (EPHFILE);
    EPHFILE = NULL; // new line, reset pointer
    free (BUFFER);
    }
    return error;
"

[0] http://aa.usno.navy.mil/software/novas/novas_faq.php